+++ lib/credit_card_model.dart
@@ -1,5 +1,6 @@
 class CreditCardModel {
-  CreditCardModel(this.cardNumber, this.expiryDate, this.cardHolderName, this.cvvCode, this.isCvvFocused);
+  CreditCardModel(this.cardNumber, this.expiryDate, this.cardHolderName,
+      this.cvvCode, this.isCvvFocused);
 
   String cardNumber = '';
   String expiryDate = '';