+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.0.0
+
+- Migrated to null safety.
+- Fixed [#17](https://github.com/SimformSolutionsPvtLtd/flutter_credit_card/issues/17) - Added translations for captions.
+- Fixed [#16](https://github.com/SimformSolutionsPvtLtd/flutter_credit_card/issues/16) - Validate Input card details.
+- Fixed [15](https://github.com/SimformSolutionsPvtLtd/flutter_credit_card/issues/15) - Change hint Color and enabled border color
+- Fixed [#9](https://github.com/SimformSolutionsPvtLtd/flutter_credit_card/issues/9) - change labelText and hintText
+- Fixed [#2](https://github.com/SimformSolutionsPvtLtd/flutter_credit_card/issues/2) - textfield validation
+
 ## 0.1.4
 *   Feature : Added number and CVV obfuscation
 *   Feature : Added input decorations